Kore.Reachability: AllPathClaim, OnePathClaim, SomeClaim
