+++ kore/src/Kore/Step/AxiomPattern.hs
@@ -16,17 +16,6 @@ import Kore.Internal.TermLike
     , VariableName
     )
 import qualified Kore.Internal.TermLike as TermLike
-import Kore.Rewriting.RewritingVariable
-    ( RewritingVariableName
-    , mkRuleVariable
-    )
-import Kore.Step.ClaimPattern
-    ( AllPathClaim
-    , OnePathClaim
-    , ReachabilityClaim (..)
-    , allPathRuleToTerm
-    , onePathRuleToTerm
-    )
 import Kore.Step.RulePattern
     ( RewriteRule
     , rewriteRuleToTerm
@@ -44,28 +33,6 @@ newtype AxiomPattern variable =
 instance Unparse (AxiomPattern VariableName) where
     unparse = unparse . getAxiomPattern
     unparse2 = unparse2 . getAxiomPattern
-
-instance From OnePathClaim (AxiomPattern VariableName) where
-    from = AxiomPattern . onePathRuleToTerm
-
-instance From OnePathClaim (AxiomPattern RewritingVariableName) where
-    from =
-        AxiomPattern
-        . TermLike.mapVariables (pure mkRuleVariable)
-        . onePathRuleToTerm
-
-instance From AllPathClaim (AxiomPattern VariableName) where
-    from = AxiomPattern . allPathRuleToTerm
-
-instance From AllPathClaim (AxiomPattern RewritingVariableName) where
-    from =
-        AxiomPattern
-        . TermLike.mapVariables (pure mkRuleVariable)
-        . allPathRuleToTerm
-
-instance From ReachabilityClaim (AxiomPattern VariableName) where
-    from (OnePath rule) = from rule
-    from (AllPath rule) = from rule
 
 instance InternalVariable variable =>
     From (RewriteRule variable) (AxiomPattern variable)
