+++ kore/test/Test/Kore/Step/RewriteStep.hs
@@ -42,10 +42,12 @@ import Kore.Internal.Predicate as Predicate
 import qualified Kore.Internal.Predicate as Predicate
 import qualified Kore.Internal.Substitution as Substitution
 import Kore.Internal.TermLike
+import Kore.Reachability
+    ( AllPathClaim (..)
+    )
 import Kore.Rewriting.RewritingVariable
 import Kore.Step.ClaimPattern
-    ( AllPathClaim (..)
-    , ClaimPattern
+    ( ClaimPattern
     , claimPattern
     , refreshExistentials
     )
